sync 19893: fix Wformat warnings (#19949)

+++ cocos/base/ccCArray.cpp
@@ -72,9 +72,9 @@ void ccArrayEnsureExtraCapacity(ccArray *arr, ssize_t extra)
 {
 	while (arr->max < arr->num + extra)
     {
-        CCLOGINFO("cocos2d: ccCArray: resizing ccArray capacity from [%d] to [%d].",
-              static_cast<int>(arr->max),
-              static_cast<int>(arr->max*2));
+        CCLOGINFO("cocos2d: ccCArray: resizing ccArray capacity from [%zd] to [%zd].",
+              arr->max,
+              arr->max*2);
 
 		ccArrayDoubleCapacity(arr);
     }
